refactor: show history using component. #2
